[FEAT] toss 연결

+++ backend/src/main/java/com/example/SKALA_Mini_Project_1/modules/bookings/service/BookingService.java
@@ -120,7 +120,7 @@ public class BookingService {
                 .bookingId(saved.getId())
                 .status(saved.getStatus())
                 .expiresAt(saved.getExpiresAt())
-                .totalPrice(saved.getTotalPrice())
+                .totalPrice(saved.getTotalPrice().longValue())
                 .seatIds(new ArrayList<>(uniqueSeatIds))
                 .build();
     }
