+++ backend/src/main/java/com/example/SKALA_Mini_Project_1/modules/payments/controller/dto/PaymentCreateRequest.java
@@ -2,8 +2,6 @@
 
 package com.example.SKALA_Mini_Project_1.modules.payments.controller.dto;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
@@ -23,6 +21,6 @@ public class PaymentCreateRequest {
     private Long seatId;
 
     @NotNull
-    private BigDecimal amount;
+    private Long amount;
 
 }
